# expensify-backend/openapi-spec.yaml
openapi: 3.0.0
paths:
...
  /users/{id}:
    patch:
      description: >-
        This endpoint allows the user to update their details. Note that this
        requires authentication of user you are updating.
      operationId: UsersController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: User update dto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDto'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: User is not authorized to update this profile
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Updates user
      tags: &ref_1
        - Users
    delete:
      description: |-
        Permanently deletes a user account and all associated data including:
            - Personal information
            - Financial records (expenses and inflows)
            - Analytics data and consent settings
            - Activity history

            This action cannot be undone. All data will be permanently erased in accordance with GDPR requirements.
      operationId: UsersController_deleteUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User and all associated data successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User and all associated data successfully deleted
                  deletedUserId:
                    type: string
                    example: user123
        '401':
          description: User is not authorized to delete this account
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Delete user account and all associated data
      tags: *ref_1
  /users/{id}/balance:
    get:
      description: Get user's current balance including total inflows and expenses
      operationId: UsersController_getCurrentBalance
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User's current balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceDto'
        '401':
          description: User is not authorized to view this balance
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Get current balance
      tags: *ref_1
  /users/monthly-balance/{year}:
    get:
      description: Get user's monthly balance breakdown for a specific year
      operationId: UsersController_getMonthlyBalance
      parameters:
        - name: year
          required: true
          in: path
          description: Year for which to get statistics
          schema:
            example: 2024
            type: number
      responses:
        '200':
          description: Monthly balance breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonthlyBalanceDto'
        '401':
          description: User is not authorized to view these statistics
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Get monthly balance
      tags: *ref_1
  /users/financial-summary:
    get:
      description: Get comprehensive financial summary including balance and statistics
      operationId: UsersController_getFinancialSummary
      parameters: []
      responses:
        '200':
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialSummaryDto'
        '401':
          description: User is not authorized to view this financial summary
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Get financial summary
      tags: *ref_1
  /users/top-tags:
    get:
      description: Get summary of top tags for both inflows and expenses
      operationId: UsersController_getTopTags
      parameters: []
      responses:
        '200':
          description: Tag summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagSummaryDto'
        '401':
          description: User is not authorized to view these tag statistics
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Get top tags
      tags: *ref_1
  /users/balance-history:
    get:
      description: >-
        Get complete history of monthly balances with cumulative totals for
        plotting trends
      operationId: UsersController_getBalanceHistory
      parameters: []
      responses:
        '200':
          description: Balance history with cumulative totals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BalanceHistoryItemDto'
        '401':
          description: User is not authorized to view this balance history
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Get balance history
      tags: *ref_1
...
  /analytics/consent:
    get:
      description: Retrieves the current cookie and tracking consent settings for the user
      operationId: AnalyticsController_getConsent
      parameters: []
      responses:
        '200':
          description: User consent settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingConsentDto'
      security: &ref_6
        - bearer: []
      summary: Get user consent settings
      tags: &ref_7
        - Analytics
    post:
      description: Updates the user's cookie and tracking consent preferences
      operationId: AnalyticsController_updateConsent
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackingConsentDto'
      responses:
        '200':
          description: Consent settings updated successfully
        '201':
          description: ''
      security: *ref_6
      summary: Update consent settings
      tags: *ref_7
  /analytics/track:
    post:
      description: Records a user activity based on consent settings
      operationId: AnalyticsController_trackActivity
      parameters: []
      requestBody:
        required: true
        description: Activity details to track
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserActivityDto'
      responses:
        '201':
          description: Activity tracked successfully
        '403':
          description: Activity tracking not allowed due to consent settings
      security: *ref_6
      summary: Track user activity
      tags: *ref_7
info:
  title: Expensify
  description: Expensify API
  version: '1.0'
  contact: {}
...
servers:
  - url: '{server}'
    description: Server URL
    variables:
      server:
        default: http://localhost:3000
        description: Server identifier
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
        updatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - username
        - firstName
        - lastName
        - email
        - updatedAt
        - createdAt
        - deletedAt
    UserUpdateDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
    BalanceDto:
      type: object
      properties:
        totalInflows:
          type: number
          description: Total amount of all inflows
          example: 5000
          minimum: 0
        totalExpenses:
          type: number
          description: Total amount of all expenses
          example: 3000
          minimum: 0
        balance:
          type: number
          description: Current balance (totalInflows - totalExpenses)
          example: 2000
      required:
        - totalInflows
        - totalExpenses
        - balance
    MonthlyBalanceDto:
      type: object
      properties:
        month:
          type: number
          description: Month number (1-12)
          example: 1
        inflow:
          type: string
          description: Total inflows for the month
          example: '1000.00'
        expense:
          type: string
          description: Total expenses for the month
          example: '800.00'
        balance:
          type: string
          description: Balance for the month
          example: '200.00'
      required:
        - month
        - inflow
        - expense
        - balance
    FinancialStatisticsDto:
      type: object
      properties:
        averageInflow:
          type: number
          description: Average amount per inflow transaction
          example: 1000
          minimum: 0
        averageExpense:
          type: number
          description: Average amount per expense transaction
          example: 800
          minimum: 0
        averageMonthlyInflow:
          type: number
          description: Average total inflows per month
          example: 1000
          minimum: 0
        averageMonthlyExpense:
          type: number
          description: Average total expenses per month
          example: 800
          minimum: 0
        totalInflowCount:
          type: number
          description: Total number of inflow transactions
          example: 25
          minimum: 0
        totalExpenseCount:
          type: number
          description: Total number of expense transactions
          example: 50
          minimum: 0
      required:
        - averageInflow
        - averageExpense
        - averageMonthlyInflow
        - averageMonthlyExpense
        - totalInflowCount
        - totalExpenseCount
    FinancialSummaryDto:
      type: object
      properties:
        currentBalance:
          description: Current balance information
          allOf:
            - $ref: '#/components/schemas/BalanceDto'
        statistics:
          description: Statistical analysis of financial transactions
          allOf:
            - $ref: '#/components/schemas/FinancialStatisticsDto'
      required:
        - currentBalance
        - statistics
    TagSummaryItemDto:
      type: object
      properties:
        tag:
          type: string
          description: Tag name
          example: groceries
        amount:
          type: number
          description: Total amount for this tag
          example: '500.00'
        type:
          type: string
          description: Type of transaction
          example: expense
          enum:
            - inflow
            - expense
      required:
        - tag
        - amount
        - type
    TagSummaryDto:
      type: object
      properties:
        inflowTags:
          description: Top inflow tags with amounts
          type: array
          items:
            $ref: '#/components/schemas/TagSummaryItemDto'
        expenseTags:
          description: Top expense tags with amounts
          type: array
          items:
            $ref: '#/components/schemas/TagSummaryItemDto'
      required:
        - inflowTags
        - expenseTags
    BalanceHistoryItemDto:
      type: object
      properties:
        year:
          type: number
          description: Year of the balance record
          example: 2024
          minimum: 1900
          maximum: 9999
        month:
          type: number
          description: Month of the balance record (1-12)
          example: 3
          minimum: 1
          maximum: 12
        inflow:
          type: string
          description: Total inflows for the period
          example: '1500.00'
        expense:
          type: string
          description: Total expenses for the period
          example: '1200.00'
        balance:
          type: string
          description: Net balance for the period (inflows - expenses)
          example: '300.00'
        cumulativeBalance:
          type: string
          description: Cumulative balance up to this period
          example: '2500.00'
      required:
        - year
        - month
        - inflow
        - expense
        - balance
        - cumulativeBalance
...
    TrackingConsentDto:
      type: object
      properties:
        analytics:
          type: boolean
          description: Consent for analytics and usage tracking
          example: false
          default: false
        social:
          type: boolean
          description: Consent for social media features and sharing
          example: false
          default: false
        advertising:
          type: boolean
          description: Consent for personalized advertising
          example: false
          default: false
      required:
        - analytics
        - social
        - advertising
    ActivityType:
      type: string
      enum:
        - AUTH
        - ERROR
        - PAGE_VIEW
        - EXPENSE_CREATED
        - BUDGET_UPDATED
        - SOCIAL_SHARE
        - SOCIAL_LIKE
        - SOCIAL_COMMENT
        - AD_INTERACTION
        - AD_CLICK
        - AD_IMPRESSION
    UserActivityDto:
      type: object
      properties:
        userId:
          type: string
          description: ID of the user performing the activity
          example: user123
        timestamp:
          format: date-time
          type: string
          description: Timestamp when the activity occurred
          example: '2024-03-15T12:00:00Z'
        action:
          description: Type of activity performed
          example: PAGE_VIEW
          allOf:
            - $ref: '#/components/schemas/ActivityType'
        metadata:
          type: object
          description: Additional data related to the activity
          example:
            page: /dashboard
            amount: 100
            category: food
          additionalProperties: true
      required:
        - userId
        - timestamp
        - action
        - metadata